Refactor categories to one component per file
